+++ Dockerfile
@@ -23,6 +23,7 @@ COPY package*.json ./
 RUN npm ci --only=production
 COPY utils/ ./utils
 COPY config/ ./config
+COPY controllers/ ./controllers
 COPY middleware/ ./middleware
 COPY routes/ ./routes
 COPY server.js ./
